pdf_converter/requirements.txt
# 基础依赖
PyQt5>=5.15.0
PyPDF2>=2.0.0
PyMuPDF>=1.18.0
pdfminer.six>=20200517
pdf2docx>=0.5.0
pdf2image>=1.16.0
Pillow>=8.0.0
tabula-py>=2.2.0
pytesseract>=0.3.8
numpy>=1.19.0
tqdm>=4.50.0

# 打包工具
pyinstaller>=4.5.0